+++ CHANGELOG.md
@@ -80,4 +80,10 @@ This will avoid infinitely looping over pages until an empty page is discovered.
 * UGCShareConnection.retrieveUGCPostsByAuthors() now includes sortBy param to get posts
 sorted by creation date.
 
-## 1.0.8 (Work in progress)
+## 1.0.8 (October 15, 2020)
+* OrganizationConnection.retrieveOrganizationBrand will return an organization brand
+ based on the organization URN.
+ OrganizationConnection.retrieveOrganizationBrandByParentOrganization will return 
+ all the organization brands linked to an orginzation  
+ 
+## 1.0.9 (Work in progress)